chore: Move change color button to be next to color picker dialog
